Refactored faulty imports for moved components in Test suites
