+++ src/test/ApparatusListBox.test.js
@@ -1,6 +1,6 @@
 import {render, screen} from "@testing-library/react"
 import React from 'react';
-import ApparatusListBox from '../components/ApparatusListBox'
+import ApparatusListBox from '../components/apparatusLists/ApparatusListBox'
 
 test('ApparatusListBox  renders without crashing', ()=>{
     render(<ApparatusListBox/>);
