+++ src/test/ApparatusSelectedListBox.test.js
@@ -1,6 +1,6 @@
 import {render, screen} from "@testing-library/react"
 import React from 'react';
-import ApparatusSelectedListBox from '../components/ApparatusSelectedListBox'
+import ApparatusSelectedListBox from '../components/apparatusLists/ApparatusSelectedListBox'
 
 test('ApparatusSelectedListBox  renders without crashing', ()=>{
     render(<ApparatusSelectedListBox/>);
