+++ src/test/WebglBox.test.js
@@ -1,6 +1,6 @@
 import {render, screen} from "@testing-library/react"
 import React from 'react';
-import WebglBox from '../components/WebglBox'
+import WebglBox from '../components/webgl/WebglBox'
 
 test('ApparatusSelectedListBox  renders without crashing', ()=>{
     render(<WebglBox/>);
